books_api/requirements.txt
Django==1.11.3
django-isbn-field==0.3
djangorestframework==3.6.3
django-extensions==1.8.1
django-filter==1.0.4
factory-boy==2.8.1
Faker==0.7.17
psycopg2==2.7.1
